requirements.txt
aiogram==3.7.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
minio==7.2.7
pydantic-settings==2.3.1
pgvector==0.2.5
google-generativeai==0.5.4 
alembic==1.13.1
apscheduler==3.10.4
dateparser==1.2.0
pytz==2024.1
pypdf==4.2.0
python-docx==1.1.2
langchain==0.2.1